+++ src/main/java/gregtech/api/task/tasks/PollutionTask.java
@@ -1,9 +1,11 @@
-package gregtech.api.task;
+package gregtech.api.task.tasks;
 
 import org.jetbrains.annotations.NotNull;
 
 import gregtech.api.enums.TickTime;
 import gregtech.api.interfaces.tileentity.IMachineProgress;
+import gregtech.api.task.TaskHost;
+import gregtech.api.task.TickableTask;
 import gregtech.common.GT_Pollution;
 
 public class PollutionTask<T extends TaskHost & IMachineProgress> extends TickableTask<T> {
